Unsets internal pointer to previously selected node on DOM deselect
```javascript
$tree.tree('selectNode');
```

When calling the above without passing a `node` parameter, the majority
of jqTree's `selectNode` method is skipped *(since no `node` has bene
provided, there is nothing to select)*.

Regardless of `node` being passed, the currently selected node is first
deselected. The deselect happens in the DOM, but the internal pointer to
the selected node is not cleared. Calling

```javascript
$tree.tree('getSelectedNode');
```

will still *(incorrectly)* return the *(now **deselected**)* node.

The small change here fixes that, causing `getSelectedNode` to properly
return `false`.