Added handlers onTreeReady and onLiAppended
It is quite useful to have handlers to manage the entire tree once it is
completed and appended to the DOM, as well a handler to manage the Li
once it is already appended to the DOM.
One useful example to these
handlers is when you want to add elements that are related to more than
one Li, for example, when creating a workflow chart.
I don't have the
knowledge (nor the time to learn) Coffee Script, so I just updated the
compiled JS file.